pyproject.toml
[project]
name = "website"
version = "0.1.0"
description = ""
authors = [{ name = "Riley Shahar", email = "riley@rileyshahar.com" }]
requires-python = ">=3.14"
dependencies = ["pyyaml (>=6.0.3,<7.0.0)"]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
